+++ CHANGELOG.md
@@ -16,6 +16,7 @@ A preview of the next release can be installed from
 - [#1604](https://github.com/babashka/babashka/issues/1604): throw `FileNotFoundException` when requiring namespace whose file cannot be found (as JVM Clojure does)
 - Bump integrant CI tests
 - [#1600](https://github.com/babashka/babashka/issues/1600): use pagesize of 64K on linux aarch64, so it works on Asahi linux
+- Expose `selmer.parser/resolve-arg`
 
 ## 1.3.182 (2023-07-20)